Day3/withoutstreamlit/requirements.txt
langchain==0.1.20
faiss-cpu==1.7.4
sentence-transformers==2.6.1
openai==1.30.1
PyMuPDF==1.23.26  # for PDF reading via fitz
google-generativeai==0.4.1
pip>=23.0
setuptools>=68.0
wheel
streamlit==1.35.0

